Refactor: 코드 리팩토링 진행
