+++ src/main/java/com/giftforyoube/global/jwt/filter/AuthenticationFilter.java
@@ -1,8 +1,9 @@
-package com.giftforyoube.global.jwt;
+package com.giftforyoube.global.jwt.filter;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.giftforyoube.global.exception.BaseResponse;
 import com.giftforyoube.global.exception.BaseResponseStatus;
+import com.giftforyoube.global.jwt.util.JwtUtil;
 import com.giftforyoube.global.security.UserDetailsImpl;
 import com.giftforyoube.user.dto.LoginRequestDto;
 import jakarta.servlet.FilterChain;
@@ -17,12 +18,12 @@ import org.springframework.security.web.authentication.UsernamePasswordAuthentic
 import java.io.IOException;
 
 @Slf4j
-public class JwtAuthenticationFilter extends UsernamePasswordAuthenticationFilter {
+public class AuthenticationFilter extends UsernamePasswordAuthenticationFilter {
 
     private final JwtUtil jwtUtil;
     private final ObjectMapper objectMapper = new ObjectMapper();
 
-    public JwtAuthenticationFilter(JwtUtil jwtUtil) {
+    public AuthenticationFilter(JwtUtil jwtUtil) {
         this.jwtUtil = jwtUtil;
         setFilterProcessesUrl("/api/login");
     }
