--- src/main/java/com/giftforyoube/user/controller/UserRedirectController.java
@@ -1,32 +1,0 @@
-package com.giftforyoube.user.controller;
-
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import java.net.URI;
-import java.net.URISyntaxException;
-
-@RestController
-@RequestMapping("/api")
-public class UserRedirectController {
-
-    // 유저 테스트 후 적용 예정
-    @GetMapping("/login/success")
-    public ResponseEntity<Void> redirectToLoginSuccessUri() throws URISyntaxException {
-        HttpHeaders headers = new HttpHeaders();
-        headers.setLocation(new URI("https://www.giftipie.me/"));
-        return new ResponseEntity<>(headers, HttpStatus.FOUND);
-    }
-
-    // 유저 테스트 후 적용 예정
-    @GetMapping("/login/fail")
-    public ResponseEntity<Void> redirectToLoginFailUri() throws URISyntaxException {
-        HttpHeaders headers = new HttpHeaders();
-        headers.setLocation(new URI("https://www.giftipie.me/"));
-        return new ResponseEntity<>(headers, HttpStatus.FOUND);
-    }
-}
