--- src/main/java/com/giftforyoube/user/dto/KakaoUserInfoDto.java
@@ -1,21 +1,0 @@
-package com.giftforyoube.user.dto;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-@JsonIgnoreProperties(ignoreUnknown = true)
-public class KakaoUserInfoDto {
-
-    private Long id;
-    private String nickname;
-    private String email;
-
-    public KakaoUserInfoDto(Long id, String nickname, String email) {
-        this.id = id;
-        this.nickname = nickname;
-        this.email = email;
-    }
-}
